Split environment config (working)
