+++ .env
@@ -1,6 +1,6 @@
 # Add Environment Variables
 
-SECRET_KEY=t.FIdc*3Y|903~"y0*_#E7?fLpDFm:-n%e-lRVqJfBQg=PB0cm'C^'9b6-LKlu533af
+SECRET_KEY=t.FIdc*3Y|903~y0*_#E7?fLpDFm:-n%e-lRVqJfBQg=PB0cmC^9b6-LKlu533af
 
 # Config DB
 DB_ENGINE=django.db.backends.postgresql_psycopg2
@@ -11,9 +11,9 @@ DB_HOST=db
 DB_PORT=5432
 
 # Config environment type (choices:  ['local', 'stage', 'prod', 'test']
-ENVIRONMENT=stage
+ENVIRONMENT=local
 
-DJANGO_DEBUG=False
+DJANGO_DEBUG=True
 
 RABBITMQ_DEFAULT_USER=admin
 RABBITMQ_DEFAULT_PASS=qwerty
@@ -22,4 +22,4 @@ HOST_NAME=mjroson.tk
 SUBDOMAIN=www
 
 PROJ_NAME=boilerplate
-IMG_VERSION=initial
+IMG_VERSION=latest
