+++ backend/config/settings.py
@@ -1,24 +1,11 @@
-"""
-Django settings for backend project.
-
-Generated by 'django-admin startproject' using Django 2.1.4.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/2.1/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/2.1/ref/settings/
-"""
-
 import os
 from apps.core.utils import get_value_env as env
 import warnings
+
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
 
-# Quick-start development settings - unsuitable for production
-# See https://docs.djangoproject.com/en/2.1/howto/deployment/checklist/
 if not env('SECRET_KEY'):
     warnings.warn((
                       "Please define SECRET_KEY before importing {0}, as a fallback "
@@ -74,7 +61,7 @@ TEMPLATES = [
     {
         'BACKEND': 'django.template.backends.django.DjangoTemplates',
         'DIRS': [
-            os.path.join(BASE_DIR, "backend/templates"),
+            os.path.join(BASE_DIR, "config/templates"),
         ],
         'APP_DIRS': True,
         'OPTIONS': {
@@ -88,8 +75,8 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'backend.wsgi.application'
-ROOT_URLCONF = 'backend.urls'
+WSGI_APPLICATION = 'config.wsgi.application'
+ROOT_URLCONF = 'config.urls'
 
 
 #############################################
