+++ backend/requirements/base.txt
@@ -1,0 +1,8 @@
+amqp==2.5.2
+celery==4.4.0
+Django==2.2
+django-celery==3.3.1
+django-celery-results==1.1.2
+django-extensions==2.2.5
+djangorestframework==3.11.0
+psycopg2==2.8.4
