+++ backend/requirements/local.txt
@@ -1,0 +1,5 @@
+-r base.txt
+
+ipdb
+graphviz==0.13.2
+pygraphviz==1.5
