+++ backend/requirements/prod.txt
@@ -1,0 +1,3 @@
+-r base.txt
+
+gunicorn
