+++ backend/requirements/test.txt
@@ -1,0 +1,4 @@
+-r base.txt
+
+pytest
+
