--- deploy_local.yml
@@ -1,67 +1,0 @@
-version: '3.1'
-services:
-  # Postgresql database
-  db:
-    image: postgres:latest
-    env_file: .env
-    container_name: ${PROJ_NAME}_db
-    environment:
-      - POSTGRES_USER=${DB_USER}
-      - POSTGRES_PASSWORD=${DB_PASS}
-      - POSTGRES_DB=${DB_NAME}
-    ports:
-      - ${DB_PORT}:${DB_PORT}
-    volumes:
-      - database:/var/lib/postgresql/data
-  
-  # RabbitMQ manager queque
-  rabbitmq:
-    image: rabbitmq
-    hostname: rabbitmq
-    env_file: .env
-    expose:
-      - "5672"  # we forward this port because it's useful for debugging
-      - "15672"  # here, we can access rabbitmq management plugin
-    volumes:
-      - rabbitmq:/var/lib/rabbitmq
-
-  # Celery worker
-  worker:
-    image: ${PROJ_NAME}_backend:${IMG_VERSION}
-    build:
-      context: ./backend
-      dockerfile: Dockerfile
-    container_name: ${PROJ_NAME}_worker
-    command: bash -c "sleep 5 && celery -A backend worker -l info --beat --scheduler django_celery_beat.schedulers:DatabaseScheduler"
-    env_file: .env
-    environment:
-      - C_FORCE_ROOT=true
-    links:
-      - db
-      - rabbitmq
-    volumes:
-      - ./backend:/usr/src/app
-    depends_on:
-      - rabbitmq
-      - db
-    
-  # Django backend
-  backend:
-    image: ${PROJ_NAME}_backend:${IMG_VERSION}
-    container_name: ${PROJ_NAME}_backend
-    command: bash -c "sleep 10 && sh run.sh"
-    env_file: .env
-    environment:
-      - ENVIRONMENT=local
-      - DJANGO_DEBUG=True
-    volumes:
-      - ./backend:/usr/src/app
-    depends_on:
-      - rabbitmq
-      - db
-    ports:
-      - "8000:8000"
-
-volumes:
-  database:
-  rabbitmq:
